// frontend/src/api/models/getTodoItemsSortBy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */

export type GetTodoItemsSortBy = typeof GetTodoItemsSortBy[keyof typeof GetTodoItemsSortBy];


export const GetTodoItemsSortBy = {
  created_at: 'created_at',
  deadline: 'deadline',
  completed_at: 'completed_at',
  updated_at: 'updated_at',
} as const;