// frontend/src/api/models/getTodoItemsSortOrder.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inspirate Consulting API
 * API for the Inspirate Consulting application
 * OpenAPI spec version: 1.0.0
 */

export type GetTodoItemsSortOrder = typeof GetTodoItemsSortOrder[keyof typeof GetTodoItemsSortOrder];


export const GetTodoItemsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;